Flesh out facade and thin out movie object
