+++ app/controllers/users/movies_controller.rb
@@ -9,15 +9,14 @@ class Users::MoviesController < ApplicationController
     @user = User.find(params[:user_id])
 
     if params[:top_movies]
-      @movies = Movie.top_movies
+      @movies = MovieFacade.top_movies
     elsif params[:movie_title]
-      @movies = Movie.search(params[:movie_title])
+      @movies = MovieFacade.search(params[:movie_title])
     end
   end
 
   def show
-    @movie = MovieFacade.new(params[:movie_id])
-    @movie = Movie.details(params[:movie_id])
+    @movie = MovieFacade.new(params[:movie_id]).movie
     @user = User.find(params[:user_id])
   end
