+++ app/views/parties/new.html.erb
@@ -1,4 +1,4 @@
-<h3><%= "Create a Movie Party for #{@movie[:title]}" %></h3>
+<h3><%= "Create a Movie Party for #{@movie.title}" %></h3>
 
 <p><%= button_to "Discover Page", "/users/#{@user.id}/discover", method: :get %></p><br>
 
@@ -6,11 +6,11 @@
 
 <h5>Viewing Party Details</h5>
 
-<%= "Movie Title: #{@movie[:title]}"%>
+<%= "Movie Title: #{@movie.title}"%>
 
-<%= form_with url: "/users/#{@user.id}/movies/#{@movie[:id]}/viewing-party", method: :post, local: true do |f| %>
+<%= form_with url: "/users/#{@user.id}/movies/#{@movie.id}/viewing-party", method: :post, local: true do |f| %>
   <%= "Duration of Party:" %>
-  <%= f.text_field :duration, value: "#{@movie[:runtime]}" %>
+  <%= f.text_field :duration, value: "#{@movie.runtime}" %>
   <%= f.label "Day:" %>
   <%= f.date_select :day %>
   <%=  "Start Time:" %>
