+++ spec/fixtures/vcr_cassettes/Show/A_user_s_show_page_should_show_the_user_s_name_a_button_to_discover_movie_and_a_list_of_viewing_parties.yml
@@ -37,7 +37,7 @@ http_interactions:
       X-Memc-Expires:
       - '21602'
       Date:
-      - Fri, 01 Dec 2023 04:44:06 GMT
+      - Fri, 01 Dec 2023 05:14:34 GMT
       Etag:
       - W/"7d53e69c5b214cb6c837bb98e4f96f14"
       Vary:
@@ -46,13 +46,13 @@ http_interactions:
       X-Cache:
       - RefreshHit from cloudfront
       Via:
-      - 1.1 f9a5caa2a18f480171fad57618db559e.cloudfront.net (CloudFront)
+      - 1.1 1152d84b916a23e4b38cef1a20064b3e.cloudfront.net (CloudFront)
       X-Amz-Cf-Pop:
       - DEN52-P2
       Alt-Svc:
       - h3=":443"; ma=86400
       X-Amz-Cf-Id:
-      - vX5tlfpMdoTMFlaHqFq7xw9RLuks1UaCWsUWC14mZOueHEcCBqOQGA==
+      - iCfm7D3NH5C4D6tS3uwLharEkSzRuU_vZH34CEGRMsnmCq7yA_Xe6A==
     body:
       encoding: ASCII-8BIT
       string: '{"adult":false,"backdrop_path":"/dQL2wJZo05GDd21VgOacMeCuyZy.jpg","belongs_to_collection":null,"budget":0,"genres":[{"id":18,"name":"Drama"},{"id":35,"name":"Comedy"},{"id":10749,"name":"Romance"},{"id":80,"name":"Crime"}],"homepage":"","id":2,"imdb_id":"tt0094675","original_language":"fi","original_title":"Ariel","overview":"After
@@ -60,7 +60,7 @@ http_interactions:
         man leaves for the city to make a living but there, he is framed and imprisoned
         for various crimes.","popularity":11.518,"poster_path":"/ojDg0PGvs6R9xYFodRct2kdI6wC.jpg","production_companies":[{"id":2303,"logo_path":null,"name":"Villealfa
         Filmproductions","origin_country":"FI"}],"production_countries":[{"iso_3166_1":"FI","name":"Finland"}],"release_date":"1988-10-21","revenue":0,"runtime":73,"spoken_languages":[{"english_name":"Finnish","iso_639_1":"fi","name":"suomi"}],"status":"Released","tagline":"","title":"Ariel","video":false,"vote_average":7.1,"vote_count":275}'
-  recorded_at: Fri, 01 Dec 2023 04:44:06 GMT
+  recorded_at: Fri, 01 Dec 2023 05:14:34 GMT
 - request:
     method: get
     uri: https://api.themoviedb.org/3/movie/2
@@ -98,7 +98,7 @@ http_interactions:
       X-Memc-Expires:
       - '21602'
       Date:
-      - Fri, 01 Dec 2023 04:44:06 GMT
+      - Fri, 01 Dec 2023 05:14:34 GMT
       Etag:
       - W/"7d53e69c5b214cb6c837bb98e4f96f14"
       Vary:
@@ -107,13 +107,13 @@ http_interactions:
       X-Cache:
       - RefreshHit from cloudfront
       Via:
-      - 1.1 422414d66021e4e123bcb2b5087f7042.cloudfront.net (CloudFront)
+      - 1.1 befcfd7ee847a3c890471f27612dbcde.cloudfront.net (CloudFront)
       X-Amz-Cf-Pop:
       - DEN52-P2
       Alt-Svc:
       - h3=":443"; ma=86400
       X-Amz-Cf-Id:
-      - onl12dR4nyF6yBEU9D3M6ZwweRHPNZ2V4VnbbfXW1EXkYufhy7w0lA==
+      - bg5fnFzu5tOIu7ZklgPbYWqQ9obtDwQtq4aVzJDWJ1qyjayDSdwhhg==
     body:
       encoding: ASCII-8BIT
       string: '{"adult":false,"backdrop_path":"/dQL2wJZo05GDd21VgOacMeCuyZy.jpg","belongs_to_collection":null,"budget":0,"genres":[{"id":18,"name":"Drama"},{"id":35,"name":"Comedy"},{"id":10749,"name":"Romance"},{"id":80,"name":"Crime"}],"homepage":"","id":2,"imdb_id":"tt0094675","original_language":"fi","original_title":"Ariel","overview":"After
@@ -121,5 +121,5 @@ http_interactions:
         man leaves for the city to make a living but there, he is framed and imprisoned
         for various crimes.","popularity":11.518,"poster_path":"/ojDg0PGvs6R9xYFodRct2kdI6wC.jpg","production_companies":[{"id":2303,"logo_path":null,"name":"Villealfa
         Filmproductions","origin_country":"FI"}],"production_countries":[{"iso_3166_1":"FI","name":"Finland"}],"release_date":"1988-10-21","revenue":0,"runtime":73,"spoken_languages":[{"english_name":"Finnish","iso_639_1":"fi","name":"suomi"}],"status":"Released","tagline":"","title":"Ariel","video":false,"vote_average":7.1,"vote_count":275}'
-  recorded_at: Fri, 01 Dec 2023 04:44:06 GMT
+  recorded_at: Fri, 01 Dec 2023 05:14:34 GMT
 recorded_with: VCR 6.2.0
